refactor: Moved hydration logic into `initHydration` so we can call it separately
